add nREPL control panel at output properties editor
