+++ .gitignore
@@ -10,3 +10,6 @@ dist/
 *~
 
 .nrepl-port
+/.calva
+/.clj-kondo
+/.lsp
