+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- Added async server interface support in `start-server!` with a client work abstraction.
+- Implemented the async Blender nREPL server directly in Basilisp.
+- Enhanced error handling to return errors at the interface layer.
+- Introduced the nREPL server control panel UI component.
+- Upgraded Basilisp to 0.2.4.
+
 ## 0.1.0
 
 - Added support for evaluating Basilisp code from Blender's Python console, file, and Text Editor.
